+++ packages/iov-encoding/src/index.ts
@@ -1,2 +1,10 @@
 export * from "./encoding";
 export { Int53, Uint32, Uint53, Uint64 } from "./integers";
+export {
+  JsonCompatibleValue,
+  isJsonCompatibleValue,
+  JsonCompatibleDictionary,
+  isJsonCompatibleDictionary,
+  JsonCompatibleArray,
+  isJsonCompatibleArray,
+} from "./json";