+++ packages/iov-encoding/types/index.d.ts
@@ -1,2 +1,3 @@
 export * from "./encoding";
 export { Int53, Uint32, Uint53, Uint64 } from "./integers";
+export { JsonCompatibleValue, isJsonCompatibleValue, JsonCompatibleDictionary, isJsonCompatibleDictionary, JsonCompatibleArray, isJsonCompatibleArray, } from "./json";